requirements.txt
Flask>=2.3.3
Office365-REST-Python-Client>=2.4.0
openai>=1.3.0
pandas
openpyxl>=3.1.2
python-dotenv>=1.0.0
Flask-WTF>=1.2.1
PyInstaller>=6.0.0